requirements.txt
langchain==0.2.0
langserve==0.0.50
crewai>=0.28
apscheduler
httpx
aiohttp
feedparser
pyyaml
web3>=6.0
eth-account
fastapi
uvicorn
sqlalchemy>=2.0
numpy
asyncio-extras 